# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CM-ANT API",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
        "/ant/api/v1/env": {
            "get": {
                "description": "Get all of the load test environments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Load Test Environment]"
                ],
                "summary": "Get the list of load test environments",
                "operationId": "LoadEnvironments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.LoadEnvRes"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ant/api/v1/load/agent": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Agent - for Development]"
                ],
                "operationId": "GetAllAgentInstallInfo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Install an agent to collect server metrics during load testing such as CPU and memory.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Agent - for Development]"
                ],
                "summary": "Install jmeter perfmon agent for metrics collection",
                "operationId": "InstallAgent",
                "parameters": [
                    {
                        "description": "agent install request",
                        "name": "loadEnvReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AgentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ant/api/v1/load/agent/{agentInstallInfoId}": {
            "delete": {
                "description": "Uninstall an agent to collect server metrics during load testing such as CPU and memory.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Agent - for Development]"
                ],
                "summary": "Uninstall jmeter perfmon agent for metrics collection",
                "operationId": "UninstallAgent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent installation info id",
                        "name": "agentInstallInfoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AgentReq": {
            "type": "object",
            "properties": {
                "pemKeyPath": {
                    "type": "string"
                },
                "publicIp": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "api.LoadEnvRes": {
            "type": "object",
            "properties": {
                "installLocation": {
                    "$ref": "#/definitions/constant.InstallLocation"
                },
                "loadEnvId": {
                    "type": "integer"
                },
                "mcisId": {
                    "type": "string"
                },
                "nsId": {
                    "type": "string"
                },
                "pemKeyPath": {
                    "type": "string"
                },
                "publicIp": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "vmId": {
                    "type": "string"
                }
            }
        },
...
        "constant.InstallLocation": {
            "type": "string",
            "enum": [
                "local",
                "remote"
            ],
            "x-enum-varnames": [
                "Local",
                "Remote"
            ]
        }
    }
}